refactor: movea auth test ui to `/authtest`
